+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 -   Go (GoLang) inference API (Beta): simple engine written in Go to do
     inference on YDF and TF-DF models.
 -   Creation of html evaluation report with plots (e.g., ROC, PR-ROC).
+-   Add support for Random Forest, CART, regressive GBT and Ranking GBT models
+    in the Go API.
 
 ## 0.2.5 - 2022-06-15